modules/person-service/requirements.txt
SQLAlchemy==1.3.19
GeoAlchemy2==0.8.4
shapely==1.7.0
aniso8601==7.0.0
attrs==19.1.0
jsonschema==3.0.2
marshmallow-sqlalchemy==0.23.1
python-dateutil==2.8.1
pytz==2020.1
psycopg2-binary==2.8.5
grpcio-tools==1.40.0
grpcio==1.40.0